requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
requests==2.32.3
